extract some common code into a base class
