+++ packages/core/src/node-execution-context/index.ts
@@ -7,3 +7,5 @@ export { PollContext } from './poll-context';
 export { SupplyDataContext } from './supply-data-context';
 export { TriggerContext } from './trigger-context';
 export { WebhookContext } from './webhook-context';
+
+export { getAdditionalKeys } from './utils';
